REFACTOR: into daily insertion and no upserts.
